Enable flexible placement of the docstring of the method __init__.
Hacked sphinx to recognise the ReST comments ".. autoinit" and "..
noautoinit" within the docstring of a class.
The comment ".. autoinit" indicates where to insert __init__'s
docstring.
The comment ".. noautoinit" suppresses the insertion of __init__'s
docstring.

This is only a quick hack, because it abuses ReST comments, but I don't
know any better solution with a comparable low effort.