Feature: Exclude selected items from autodoc.
This commit hacks autodoc to recognize the special ReST-comment
".. no-autodoc" within Python doc-strings. If a
docstring contains the comment autodoc ignores this item.

This commit replaces a similar implementation, that used the 
'autodoc-skip-member' hook.

Beware: this is a hack, it misuses ReST-comments. Unfortunately I don't
know a better solution with a comparable low effort.